Move to IKVM.Util. Technically not an extern. Add comments. Move polyfill implementations into one class with calls to real implementations to clean up call sites. Remove some inheritence. Seemed to be useful to reference static properties from the base class. But ultimately these are all just static classes. Revert from file-scoped namespaces. I like these, but the whole project doens't use them. Better to keep it consistent for now.
