+++ src/IKVM.Java/map.xml
@@ -1648,7 +1648,7 @@
                     <ldarg_0 />
                     <ldfld class="com.sun.crypto.provider.AESCrypt" name="K" sig="[I" />
                     <!-- bool AESCrypt.DecryptBlock(byte[] in, int inOffset, byte[] out, int outOffset, int[] K) -->
-                    <call type="IKVM.Java.Externs.com.sun.crypto.provider.AESCrypt" name="DecryptBlock" sig="([BI[BI[I)Z" />
+                    <call type="IKVM.Runtime.Util.Com.Sun.Crypto.Provider.AESCrypt" name="DecryptBlock" sig="([BI[BI[I)Z" />
                     <brfalse name="continue" />
                     <ret />
                     <label name="continue" />
@@ -1663,7 +1663,7 @@
                     <ldarg_0 />
                     <ldfld class="com.sun.crypto.provider.AESCrypt" name="K" sig="[I" />
                     <!-- bool AESCrypt.EncryptBlock(byte[] in, int inOffset, byte[] out, int outOffset, int[] K) -->
-                    <call type="IKVM.Java.Externs.com.sun.crypto.provider.AESCrypt" name="EncryptBlock" sig="([BI[BI[I)Z" />
+                    <call type="IKVM.Runtime.Util.Com.Sun.Crypto.Provider.AESCrypt" name="EncryptBlock" sig="([BI[BI[I)Z" />
                     <brfalse name="continue" />
                     <ret />
                     <label name="continue" />
@@ -1679,7 +1679,7 @@
                     <ldarg_3 />
                     <ldarg_s argNum="4" />
                     <!-- bool GHASH.ProcessBlocks(byte[] data, int inOffs, int blocks, long[] state, long[] subH) -->
-                    <call type="IKVM.Java.Externs.com.sun.crypto.provider.GHASH" name="ProcessBlocks" sig="([BII[J[J)Z" />
+                    <call type="IKVM.Runtime.Util.Com.Sun.Crypto.Provider.GHASH" name="ProcessBlocks" sig="([BII[J[J)Z" />
                     <brfalse name="continue" />
                     <ret />
                     <label name="continue" />
@@ -1696,7 +1696,7 @@
                     <ldarg_s argNum="4" />
                     <ldarg_s argNum="5" />
                     <!-- bool CipherBlockChaining.DecryptAESCrypt(object self, byte[] cipher, int cipherOffset, int cipherLen, byte[] plain, int plainOffset, ) -->
-                    <call type="IKVM.Java.Externs.com.sun.crypto.provider.CipherBlockChaining" name="DecryptAESCrypt" sig="(Ljava.lang.Object;[BII[BI)Z" />
+                    <call type="IKVM.Runtime.Util.Com.Sun.Crypto.Provider.CipherBlockChaining" name="DecryptAESCrypt" sig="(Ljava.lang.Object;[BII[BI)Z" />
                     <brfalse name="continue" />
                     <ldarg_3 />
                     <ret />
@@ -1712,7 +1712,7 @@
                     <ldarg_s argNum="4" />
                     <ldarg_s argNum="5" />
                     <!-- bool CipherBlockChaining.EncryptAESCrypt(object self, byte[] plain, int plainOffset, int plainLen, byte[] cipher, int cipherOffset) -->
-                    <call type="IKVM.Java.Externs.com.sun.crypto.provider.CipherBlockChaining" name="EncryptAESCrypt" sig="(Ljava.lang.Object;[BII[BI)Z" />
+                    <call type="IKVM.Runtime.Util.Com.Sun.Crypto.Provider.CipherBlockChaining" name="EncryptAESCrypt" sig="(Ljava.lang.Object;[BII[BI)Z" />
                     <brfalse name="continue" />
                     <ldarg_3 />
                     <ret />
